+++ src/main/java/com/recipe/app/src/fridgeBasket/api/FridgeBasketController.java
@@ -5,6 +5,7 @@ import com.recipe.app.src.fridgeBasket.application.FridgeBasketService;
 import com.recipe.app.src.fridgeBasket.application.dto.FridgeBasketDto;
 import com.recipe.app.src.user.domain.SecurityUser;
 import com.recipe.app.src.user.domain.User;
+import com.recipe.app.src.user.exception.UserTokenNotExistException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
@@ -21,6 +22,9 @@ public class FridgeBasketController {
     @ResponseBody
     @PostMapping("/basket")
     public BaseResponse<Void> postFridgesBasket(final Authentication authentication, @RequestBody FridgeBasketDto.FridgeBasketIdsRequest request) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         fridgeBasketService.createFridgesBasket(request, user);
 
@@ -30,6 +34,9 @@ public class FridgeBasketController {
     @ResponseBody
     @PostMapping("/direct-basket")
     public BaseResponse<FridgeBasketDto.DirectFridgeBasketsResponse> postFridgesDirectBasket(final Authentication authentication, @RequestBody FridgeBasketDto.DirectFridgeBasketsRequest request) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         FridgeBasketDto.DirectFridgeBasketsResponse data = new FridgeBasketDto.DirectFridgeBasketsResponse(fridgeBasketService.createDirectFridgeBasket(request, user));
 
@@ -38,6 +45,9 @@ public class FridgeBasketController {
 
     @GetMapping("/basket")
     public BaseResponse<FridgeBasketDto.FridgeBasketsResponse> getFridgesBasket(final Authentication authentication) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         FridgeBasketDto.FridgeBasketsResponse data = new FridgeBasketDto.FridgeBasketsResponse(fridgeBasketService.countFridgeBasketsByUser(user), fridgeBasketService.retrieveFridgeBasketsByUser(user));
 
@@ -46,6 +56,9 @@ public class FridgeBasketController {
 
     @DeleteMapping("/basket")
     public BaseResponse<Void> deleteFridgeBasket(final Authentication authentication, @RequestParam(value = "ingredient") String ingredient) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         fridgeBasketService.deleteFridgeBasket(user, ingredient);
 
@@ -55,6 +68,9 @@ public class FridgeBasketController {
     @ResponseBody
     @PatchMapping("/basket")
     public BaseResponse<Void> patchFridgeBaskets(final Authentication authentication, @RequestBody FridgeBasketDto.FridgeBasketsRequest request) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         fridgeBasketService.updateFridgeBaskets(request, user);
 
@@ -63,6 +79,9 @@ public class FridgeBasketController {
 
     @GetMapping("/basket/count")
     public BaseResponse<FridgeBasketDto.FridgeBasketsCountResponse> getFridgesBasketCount(final Authentication authentication) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         FridgeBasketDto.FridgeBasketsCountResponse data = new FridgeBasketDto.FridgeBasketsCountResponse(fridgeBasketService.countFridgeBasketsByUser(user));
