+++ src/main/java/com/recipe/app/src/userRecipe/api/UserRecipeController.java
@@ -3,6 +3,7 @@ package com.recipe.app.src.userRecipe.api;
 import com.recipe.app.common.response.BaseResponse;
 import com.recipe.app.src.user.domain.SecurityUser;
 import com.recipe.app.src.user.domain.User;
+import com.recipe.app.src.user.exception.UserTokenNotExistException;
 import com.recipe.app.src.userRecipe.application.UserRecipeService;
 import com.recipe.app.src.userRecipe.application.dto.UserRecipeDto;
 import lombok.RequiredArgsConstructor;
@@ -23,6 +24,9 @@ public class UserRecipeController {
 
     @GetMapping("")
     public BaseResponse<List<UserRecipeDto.UserRecipesResponse>> getUserRecipes(final Authentication authentication) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         List<UserRecipeDto.UserRecipesResponse> data = userRecipeService.retrieveUserRecipes(user).stream()
                 .map(UserRecipeDto.UserRecipesResponse::new)
@@ -33,6 +37,9 @@ public class UserRecipeController {
 
     @GetMapping("/{userRecipeIdx}")
     public BaseResponse<UserRecipeDto.UserRecipeResponse> getUserRecipe(final Authentication authentication, @PathVariable int userRecipeIdx) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         UserRecipeDto.UserRecipeResponse data = new UserRecipeDto.UserRecipeResponse(userRecipeService.retrieveUserRecipe(user, userRecipeIdx));
 
@@ -42,6 +49,9 @@ public class UserRecipeController {
     @ResponseBody
     @PostMapping("")
     public BaseResponse<UserRecipeDto.UserRecipeResponse> postUserRecipe(final Authentication authentication, @RequestBody UserRecipeDto.UserRecipeRequest request) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         UserRecipeDto.UserRecipeResponse data = new UserRecipeDto.UserRecipeResponse(userRecipeService.createUserRecipe(user, request));
 
@@ -51,6 +61,9 @@ public class UserRecipeController {
     @ResponseBody
     @PatchMapping("/{userRecipeIdx}")
     public BaseResponse<UserRecipeDto.UserRecipeResponse> patchUserRecipe(final Authentication authentication, @PathVariable int userRecipeIdx, @RequestBody UserRecipeDto.UserRecipeRequest request) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         UserRecipeDto.UserRecipeResponse data = new UserRecipeDto.UserRecipeResponse(userRecipeService.updateUserRecipe(user, userRecipeIdx, request));
 
@@ -59,6 +72,9 @@ public class UserRecipeController {
 
     @DeleteMapping("/{userRecipeIdx}")
     public BaseResponse<Void> deleteUserRecipe(final Authentication authentication, @PathVariable int userRecipeIdx) {
+        if (authentication == null)
+            throw new UserTokenNotExistException();
+
         User user = ((SecurityUser) authentication.getPrincipal()).getUser();
         userRecipeService.deleteUserRecipe(user, userRecipeIdx);
